+++ test/spec/api/pointy-core.spec.ts
@@ -4,6 +4,9 @@ import { ExampleUser } from '../../../src/models';
 import { ExampleOwner } from '../../examples/api/models/example-owner';
 import { ExampleRelation } from '../../examples/api/models/example-relation';
 import { HookTestClass } from '../../examples/api/models/hook-test-class';
+import { setLog } from '../../../src/log';
+import { Log } from 'ts-tiny-log';
+import { LogLevel } from 'ts-tiny-log/levels';
 
 let ipcMessage;
 let errorHandler;
@@ -36,6 +39,9 @@ beforeAll(async () => {
 
 	jasmine.DEFAULT_TIMEOUT_INTERVAL = 15000;
 	process.env.PORT = '8081';
+
+	// Disable logging
+	setLog(new Log({ level: LogLevel.none }));
 });
 
 describe('Pointy Core', () => {