+++ test/spec/api/utils/run-hook.spec.ts
@@ -40,17 +40,6 @@ class TestHooks extends BaseModel {
  * pointyapi/utils
  */
 describe('[Utils] runHook()', () => {
-	let cwarn;
-
-	beforeEach(() => {
-		cwarn = console.warn;
-		console.warn = () => {};
-	});
-
-	afterEach(() => {
-		console.warn = cwarn;
-	});
-
 	it('runs the hook', async () => {
 		const { request, response } = createMockRequest();